+++ web/src/__generated__/apis/CommerceApi.ts
@@ -147,6 +147,10 @@ export class CommerceApi {
   }
   
   /**
+   * @remarks
+   * **Authentication:**
+   * This method requires a valid bearer token in the `Authorization` header.
+   * 
    * @param {bigint | string} objectId - Gamertag of the player.Underlying objectId type is integer in format int64.
    * @param {string} gamertag - Override the Gamer Tag of the player. This is generally inferred by the auth token.
    * @returns {Promise<HttpResponse<GetTotalCouponResponse>>} A promise containing the HttpResponse of GetTotalCouponResponse
@@ -159,7 +163,8 @@ export class CommerceApi {
       r: this.r,
       e,
       m: GET,
-      g: gamertag
+      g: gamertag,
+      w: true
     });
   }