+++ web/src/__generated__/apis/EventPlayersApi.ts
@@ -65,6 +65,10 @@ export class EventPlayersApi {
   }
   
   /**
+   * @remarks
+   * **Authentication:**
+   * This method requires a valid bearer token in the `Authorization` header.
+   * 
    * @param {EventScoreRequest} payload - The `EventScoreRequest` instance to use for the API request
    * @param {bigint | string} objectId - Gamertag of the player.Underlying objectId type is integer in format int64.
    * @param {string} gamertag - Override the Gamer Tag of the player. This is generally inferred by the auth token.
@@ -79,7 +83,8 @@ export class EventPlayersApi {
       e,
       m: PUT,
       p: payload,
-      g: gamertag
+      g: gamertag,
+      w: true
     });
   }
 }