+++ web/src/__generated__/apis/TournamentsApi.ts
@@ -442,6 +442,10 @@ export class TournamentsApi {
   }
   
   /**
+   * @remarks
+   * **Authentication:**
+   * This method requires a valid bearer token in the `Authorization` header.
+   * 
    * @param {bigint | string} objectId - Gamertag of the player.Underlying objectId type is integer in format int64.
    * @param {string} gamertag - Override the Gamer Tag of the player. This is generally inferred by the auth token.
    * @returns {Promise<HttpResponse<TournamentClientView>>} A promise containing the HttpResponse of TournamentClientView
@@ -454,7 +458,8 @@ export class TournamentsApi {
       r: this.r,
       e,
       m: GET,
-      g: gamertag
+      g: gamertag,
+      w: true
     });
   }
 }